Refactor creating Lisk account
